# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Libre-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/studies/{slug}": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study",
        "description": "Get study details.",
        "operationId": "get_study_api_admin_studies__slug__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Update Study",
        "description": "Update study configuration (draft only).",
        "operationId": "update_study_api_admin_studies__slug__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Delete Study",
        "description": "Delete a study (Superuser only, and must be Archived).",
        "operationId": "delete_study_api_admin_studies__slug__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/studies/{slug}/stats": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study Stats",
        "description": "Get aggregated study statistics.",
        "operationId": "get_study_stats_api_admin_studies__slug__stats_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyStatsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/export/config": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Export Study Config",
        "description": "Export study configuration without participant data.\nReturns clean JSON suitable for import.",
        "operationId": "export_study_config_api_admin_studies__slug__export_config_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/validate-import": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Validate Study Import",
        "description": "Validate imported configuration without creating study.\nReturns validation results and warnings.",
        "operationId": "validate_study_import_api_admin_studies_validate_import_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Config"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/import": {
      "post": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Import Study Config",
        "description": "Import study configuration from exported JSON.\nCreates a new study in draft state.",
        "operationId": "import_study_config_api_admin_studies_import_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Project-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/storage-usage": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Study Storage Usage",
        "description": "Get audio storage usage statistics for a study.\n\nReturns:\n    Dictionary with total_bytes, total_mb, file_count, quota_mb, quota_bytes, usage_percent",
        "operationId": "get_study_storage_usage_api_admin_studies__slug__storage_usage_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/studies/{slug}/participants/{participant_id}/export/audio": {
      "get": {
        "tags": [
          "admin-exports",
          "Admin Exports"
        ],
        "summary": "Export Participant Audio",
        "description": "Export all audio recordings for a participant as a ZIP with metadata.",
        "operationId": "export_participant_audio_api_admin_studies__slug__participants__participant_id__export_audio_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audio/{recording_id}": {
      "delete": {
        "tags": [
          "audio"
        ],
        "summary": "Delete Audio Recording",
        "description": "Delete an audio recording (before submission only).\n\nArgs:\n    recording_id: ID of the recording to delete\n    session_token: Participant session token (query parameter)\n\nArgs:\n    recording_id: ID of recording to delete\n    session_token: Participant session token for authorization\n\nReturns:\n    Success message\n\nRaises:\n    HTTPException: If not authorized or already submitted",
        "operationId": "delete_audio_recording_api_audio__recording_id__delete",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Recording Id"
            }
          },
          {
            "name": "session_token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StudyImportRequest": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "new_slug": {
            "type": "string",
            "title": "New Slug"
          }
        },
        "type": "object",
        "required": [
          "config",
          "new_slug"
        ],
        "title": "StudyImportRequest"
      },
...
      "StudyStatsRead": {
        "properties": {
          "started_count": {
            "type": "integer",
            "title": "Started Count"
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count"
          },
          "completion_rate": {
            "type": "number",
            "title": "Completion Rate"
          },
          "median_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Duration Seconds"
          },
          "device_breakdown": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Device Breakdown"
          }
        },
        "type": "object",
        "required": [
          "started_count",
          "completed_count",
          "completion_rate",
          "median_duration_seconds",
          "device_breakdown"
        ],
        "title": "StudyStatsRead",
        "description": "Schema for study statistics."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}